+++ addOns/authhelper/src/test/java/org/zaproxy/addon/authhelper/AuthenticationDetectionScanRuleUnitTest.java
@@ -303,6 +303,36 @@ class AuthenticationDetectionScanRuleUnitTest
     }
 
     @Test
+    void shouldAlertOnPostWithJsonArrayAuthParams() throws HttpMalformedHeaderException {
+        // Given
+        HttpMessage msg = new HttpMessage();
+        msg.setRequestHeader("POST http://www.example.com/login/ HTTP/1.1");
+        msg.getRequestHeader().setHeader(HttpHeader.CONTENT_TYPE, HttpHeader.JSON_CONTENT_TYPE);
+        msg.setRequestBody("[{\"email\":\"test@test.com\",\"password\":\"test123\"}]");
+        msg.getRequestHeader().setContentLength(msg.getRequestBody().length());
+        msg.setResponseBody("<html></html>");
+        msg.setResponseHeader(
+                "HTTP/1.1 200 OK\r\n"
+                        + "Server: Apache-Coyote/1.1\r\n"
+                        + "Content-Type: text/html;charset=ISO-8859-1\r\n"
+                        + "Content-Length: "
+                        + msg.getResponseBody().length()
+                        + "\r\n");
+        // When
+        scanHttpResponseReceive(msg);
+        // Then
+        assertThat(alertsRaised.size(), equalTo(1));
+        assertThat(alertsRaised.get(0).getName(), equalTo("Authentication Request Identified"));
+        assertThat(alertsRaised.get(0).getConfidence(), equalTo(3));
+        assertThat(alertsRaised.get(0).getParam(), equalTo("[0].email"));
+        assertThat(alertsRaised.get(0).getEvidence(), equalTo("[0].password"));
+        assertThat(
+                alertsRaised.get(0).getOtherInfo(),
+                equalTo(
+                        "userParam=[0].email\nuserValue=test@test.com\npasswordParam=[0].password"));
+    }
+
+    @Test
     void shouldAlertOnPostWithDeeperJsonAuthParams() throws HttpMalformedHeaderException {
         // Given
         HttpMessage msg = new HttpMessage();